Finalize renaming
